[microvm] E: Rename Memory Thread Module
