+++ src/microvm/src/memory_thread.rs
